+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- The stub will remove the euconsent cookie if the cmpId is set to 1.
 
 ## ~~[2.0.1-beta.1] - 2019-09-13~~
 ### ~~Added~~